Update README.md
To build a file-based key-value data store that supports the basic CRD operations C-Create, R-Read, D-Delete. This data store is meant to be used as a local storage for one single process on one laptop. the data store must be exposed as a library to clients that can instantiate a class and work with the data store. It supports both functional and non-functional requirements.